refactored http request generator
