+++ api/exec.go
@@ -7,7 +7,6 @@ import (
 	"github.com/saeidalz13/gurl/internal/appconstants"
 	"github.com/saeidalz13/gurl/internal/domainparser"
 	"github.com/saeidalz13/gurl/internal/errutils"
-	"github.com/saeidalz13/gurl/internal/httpconstants"
 	"github.com/saeidalz13/gurl/internal/methodparser"
 	"github.com/saeidalz13/gurl/internal/terminalutils"
 )
@@ -36,15 +35,13 @@ func ExecGurl() {
 		return
 	}
 
-	contentType := determineContentType(cp.DataType)
-
-	httpRequest := NewHTTPRequestGenerator(
+	httpRequest := http.NewHTTPRequestGenerator(
 		dp.Domain,
 		dp.Path,
 		cp.Cookies,
 		method,
-		contentType,
 		cp.Data,
+		cp.DataType,
 	).Generate()
 
 	if cp.Verbose {
@@ -67,17 +64,4 @@ func manageWebSocket(dp domainparser.DomainParser, tcm tcp.TCPConnManager, verbo
 
 	go tcm.ReadWebSocketData(secWsKey, verbose)
 	tcm.WriteWebSocketData([]byte(wsRequest))
-}
-
-func determineContentType(dataType uint8) string {
-	switch dataType {
-	case httpconstants.DataTypeJson:
-		return "application/json"
-	case httpconstants.DataTypeText:
-		return "text/plain"
-	case httpconstants.DataTypeImage:
-		return "image/jpeg"
-	}
-
-	return ""
 }
